LOPO-6: Have localisation correctly process observations of unique goalposts

+++ shared/messages/localisation/FieldObject.h
@@ -5,14 +5,14 @@ namespace messages {
     namespace localisation {
         struct FieldObject {
             std::string name;
-            float wm_x;
-            float wm_y;
-            float heading;
-            float sd_x;
-            float sd_y;
-            float sr_xx;
-            float sr_xy;
-            float sr_yy;
+            double wm_x;
+            double wm_y;
+            double heading;
+            double sd_x;
+            double sd_y;
+            double sr_xx;
+            double sr_xy;
+            double sr_yy;
             bool lost;
         };
     }
